requirements.txt
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.2.0
django-filter==22.1
openpyxl==3.0.10
#mysqlclient==2.1.1
#psycopg2==2.9.5
drf-extra-fields==3.4.1
lxml==4.9.1
django-ordered-model==3.6
django-utils-six #token, don't important
djoser #gữi mail
knox #login
cryptography #knox
jinja2 #knox
django-rest-knox
boto3
django-simple-history
dj-database-url==0.5.0
gunicorn==20.0.4

pytz==2020.1
dj-static